Ditch geometry server dependency

+++ backend/agent/tools/code_executor_tool.py
@@ -3,8 +3,10 @@ import docker
 import tempfile
 import os
 import pathlib
+import uuid
+from datetime import datetime
+from chainlit.data.storage_clients.s3 import S3StorageClient
 
-import requests
 
 @tool
 def code_executor_tool(code: str) -> str:
@@ -58,17 +60,50 @@ def code_executor_tool(code: str) -> str:
     container.remove()
 
     vtp_files = [f for f in os.listdir(output_dir) if f.endswith('.vtp')]
+    if not vtp_files:
+        return "Error: No .vtp file generated."
+
     vtp_file_name = vtp_files[0]
     vtp_file_path = os.path.join(output_dir, vtp_file_name)
 
     os.unlink(temp_file)
 
-    with open(vtp_file_path, 'rb') as f:
-        files = {'file': (vtp_file_name, f, 'application/octet-stream')}
-        response = requests.post(os.getenv('BACKEND_HOST')+'/upload', files=files)
-        upload_data = response.json()
-        s3_key = upload_data.get("s3_key")
-        print(f"/files/by-key/{s3_key}")
+    # S3 configuration
+    AWS_REGION = os.getenv("APP_AWS_REGION", "eu-north-1")
+    AWS_ACCESS_KEY = os.getenv("APP_AWS_ACCESS_KEY")
+    AWS_SECRET_KEY = os.getenv("APP_AWS_SECRET_KEY")
+    BUCKET_NAME = os.getenv("BUCKET_NAME", "keystone-user-content-files")
 
-        return f"/files/by-key/{s3_key}"
+    if not all([AWS_ACCESS_KEY, AWS_SECRET_KEY, BUCKET_NAME]):
+        return "Error: Missing AWS configuration for file upload."
+
+    storage_client = S3StorageClient(
+        bucket=BUCKET_NAME,
+        region_name=AWS_REGION,
+        aws_access_key_id=AWS_ACCESS_KEY,
+        aws_secret_access_key=AWS_SECRET_KEY,
+    )
+
+    with open(vtp_file_path, 'rb') as f:
+        vtp_file_content = f.read()
+
+    file_extension = vtp_file_name.lower().split(".")[-1]
+    unique_id = str(uuid.uuid4())
+    year = datetime.now().year
+    month = datetime.now().strftime("%m")
+
+    s3_key = f"user-uploads/{year}/{month}/{unique_id}.{file_extension}"
+
+    upload_data = storage_client.sync_upload_file(
+        object_key=s3_key,
+        data=vtp_file_content,
+        mime='application/octet-stream'
+    )
+
+    s3_key_from_upload = upload_data.get("object_key")
+    if s3_key_from_upload:
+        print(f"/files/by-key/{s3_key_from_upload}")
+        return f"/files/by-key/{s3_key_from_upload}"
+    else:
+        return "Error: Failed to upload file to S3."
